refactor(dto): use BigDecimal for amounts, TransactionType enum and records for DTOs

+++ src/main/java/com/fintrack/fintrack_backend/dto/DashboardResponse.java
@@ -1,37 +1,38 @@
 package com.fintrack.fintrack_backend.dto;
+import java.math.BigDecimal;
 
 public class DashboardResponse {
-    private double income;
-    private double expense;
-    private double balance;
+    private BigDecimal income;
+    private BigDecimal expense;
+    private BigDecimal balance;
 
-    public DashboardResponse(double balance, double income, double expense) {
+    public DashboardResponse(BigDecimal balance, BigDecimal income, BigDecimal expense) {
         this.balance = balance;
         this.income = income;
         this.expense = expense;
     }
 
-    public double getIncome() {
+    public BigDecimal getIncome() {
         return income;
     }
 
-    public void setIncome(double income) {
+    public void setIncome(BigDecimal income) {
         this.income = income;
     }
 
-    public double getExpense(){
+    public BigDecimal getExpense(){
         return expense;
     }
 
-    public void setExpense(double expense){
+    public void setExpense(BigDecimal expense){
         this.expense = expense;
     }
 
-    public double getBalance() {
+    public BigDecimal getBalance() {
         return balance;
     }
 
-    public void setBalance(double balance) {
+    public void setBalance(BigDecimal balance) {
         this.balance = balance;
     }
